refactor steam callback router into a handler class
